internals: do not use (kill-this-buffer)
The reason as to why we should not be using (kill-this-buffer) is discussed
here: https://www.reddit.com/r/emacs/comments/64xb3q/killthisbuffer_sometimes_just_stops_working/

Essentially, `kill-this-buffer' is meant to be used in conjunction with the
menu bar, and for some reason stops working when a menu bar isn't present,
which is our case.

This commit fixes all these issues.